72-2 Completing Get Me And My Profile Page With Caching

+++ src/proxy.ts
@@ -1,61 +1,70 @@
-
-import { NextResponse } from 'next/server'
-import type { NextRequest } from 'next/server'
 import jwt, { JwtPayload } from 'jsonwebtoken';
-// import { cookies } from 'next/headers';
+import type { NextRequest } from 'next/server';
+import { NextResponse } from 'next/server';
 import { getDefaultDashboardRoute, getRouteOwner, isAuthRoute, UserRole } from './lib/auth-utils';
-import { deleteCookie, getCookie } from './services/auth/tokenHandler';
-
+import { getUserInfo } from './services/auth/getUserInfo';
+import { deleteCookie, getCookie } from './services/auth/tokenHandlers';
+import { getNewAccessToken } from './services/auth/auth.service';
 
 
 
 // This function can be marked `async` if using `await` inside
 export async function proxy(request: NextRequest) {
-  // const cookieStore = await cookies()
+    const pathname = request.nextUrl.pathname;
+    const hasTokenRefreshedParam = request.nextUrl.searchParams.has('tokenRefreshed');
 
-  console.log("pathname", request.nextUrl.pathname)
-  const pathname = request.nextUrl.pathname;
-
-  // const accessToken = request.cookies.get("accessToken")?.value || null;
-
-  const accessToken= await getCookie("accessToken") || null;
-
-
-  let userRole: string | null = null;
-
-  if (accessToken) {
-    const verifiedToken: JwtPayload | string = jwt.verify(accessToken, process.env.JWT_SECRET as string);
-    console.log(verifiedToken)
-
-    if (typeof verifiedToken === "string") {
-      // cookieStore.delete("accessToken");
-      // cookieStore.delete("refreshToken");
-      await deleteCookie("accessToken");
-      await deleteCookie("refreshToken");
-
-      return NextResponse.redirect(new URL('/login', request.url));
+    // If coming back after token refresh, remove the param and continue
+    if (hasTokenRefreshedParam) {
+        const url = request.nextUrl.clone();
+        url.searchParams.delete('tokenRefreshed');
+        return NextResponse.redirect(url);
     }
-    userRole = verifiedToken.role
-  }
 
-  const routeOwner = getRouteOwner(pathname);
-  // path = /doctor/appointment => DOCTOR
-  // path = /my-profile => COMMON
-  // path = /login => null
+    const tokenRefreshResult = await getNewAccessToken();
 
-  const isAuth = isAuthRoute(pathname); // true | false
+    // If token was refreshed, redirect to same page to fetch with new token
+    if (tokenRefreshResult?.tokenRefreshed) {
+        const url = request.nextUrl.clone();
+        url.searchParams.set('tokenRefreshed', 'true');
+        return NextResponse.redirect(url);
+    }
 
-  // rule-1  : user logged in and trying to access auth route => redirect to dashboard
-  if (accessToken && isAuth) {
-    return NextResponse.redirect(new URL(getDefaultDashboardRoute(userRole as UserRole), request.url));
-  }
+    // const accessToken = request.cookies.get("accessToken")?.value || null;
 
-  //  rule-2 : user not logged in and trying to access open public route
-  if (routeOwner === null) {
-    return NextResponse.next()
-  }
+    const accessToken = await getCookie("accessToken") || null;
 
-  //  rule-1 and rule-2 for public route and auth routes 
+    let userRole: UserRole | null = null;
+    if (accessToken) {
+        const verifiedToken: JwtPayload | string = jwt.verify(accessToken, process.env.JWT_SECRET as string);
+
+        if (typeof verifiedToken === "string") {
+            await deleteCookie("accessToken");
+            await deleteCookie("refreshToken");
+            return NextResponse.redirect(new URL('/login', request.url));
+        }
+
+        userRole = verifiedToken.role;
+    }
+
+    const routerOwner = getRouteOwner(pathname);
+    //path = /doctor/appointments => "DOCTOR"
+    //path = /my-profile => "COMMON"
+    //path = /login => null
+
+    const isAuth = isAuthRoute(pathname)
+
+    // Rule 1 : User is logged in and trying to access auth route. Redirect to default dashboard
+    if (accessToken && isAuth) {
+        return NextResponse.redirect(new URL(getDefaultDashboardRoute(userRole as UserRole), request.url))
+    }
+
+
+    // Rule 2 : User is trying to access open public route
+    if (routerOwner === null) {
+        return NextResponse.next();
+    }
+
+    // Rule 1 & 2 for open public routes and auth routes
 
     if (!accessToken) {
         const loginUrl = new URL("/login", request.url);
@@ -63,37 +72,50 @@ export async function proxy(request: NextRequest) {
         return NextResponse.redirect(loginUrl);
     }
 
+    // Rule 3 : User need password change
 
-  // rule-3 : user and trying to access protected route
-  if (routeOwner === "COMMON") {
-    return NextResponse.next()
-  }
+    if (accessToken) {
+        const userInfo = await getUserInfo();
+        if (userInfo.needPasswordChange) {
+            if (pathname !== "/reset-password") {
+                const resetPasswordUrl = new URL("/reset-password", request.url);
+                resetPasswordUrl.searchParams.set("redirect", pathname);
+                return NextResponse.redirect(resetPasswordUrl);
+            }
+            return NextResponse.next();
+        }
 
-  // rule-4 : user trying to access role based protected route
-
-  if (routeOwner === "ADMIN" || routeOwner === "DOCTOR" || routeOwner === "PATIENT") {
-    if (userRole !== routeOwner) {
-      return NextResponse.redirect(new URL(getDefaultDashboardRoute(userRole as UserRole), request.url));
+        if (userInfo && !userInfo.needPasswordChange && pathname === '/reset-password') {
+            return NextResponse.redirect(new URL(getDefaultDashboardRoute(userRole as UserRole), request.url));
+        }
     }
-    return NextResponse.next()
-  }
 
+    // Rule 4 : User is trying to access common protected route
+    if (routerOwner === "COMMON") {
+        return NextResponse.next();
+    }
 
-  return NextResponse.next()
+    // Rule 5 : User is trying to access role based protected route
+    if (routerOwner === "ADMIN" || routerOwner === "DOCTOR" || routerOwner === "PATIENT") {
+        if (userRole !== routerOwner) {
+            return NextResponse.redirect(new URL(getDefaultDashboardRoute(userRole as UserRole), request.url))
+        }
+    }
+
+    return NextResponse.next();
 }
 
 
 
-// used negative matcher for this 
 export const config = {
-  matcher: [
-    /*
-     * Match all request paths except for the ones starting with:
-     * - api (API routes)
-     * - _next/static (static files)
-     * - _next/image (image optimization files)
-     * - favicon.ico, sitemap.xml, robots.txt (metadata files)
-     */
-    '/((?!api|_next/static|_next/image|favicon.ico|sitemap.xml|robots.txt|.well-known).*)',
-  ],
+    matcher: [
+        /*
+         * Match all request paths except for the ones starting with:
+         * - api (API routes)
+         * - _next/static (static files)
+         * - _next/image (image optimization files)
+         * - favicon.ico, sitemap.xml, robots.txt (metadata files)
+         */
+        '/((?!api|_next/static|_next/image|favicon.ico|sitemap.xml|robots.txt|.well-known).*)',
+    ],
 }
